Corrected implementation of RM and EDF queues

+++ rclcpp/include/rclcpp/experimental/executors/events_executor/deadline_events_queue.hpp
@@ -64,10 +64,12 @@ public:
         PriorityEvent priority_event = {deadline, single_event};
         for (size_t ev = 0; ev < event.num_events; ev++) {
           timers_queue_.push(priority_event);
+          // std::cout << "Enqueued timer of deadline " << deadline << std::endl;
         }
       } else {
         for (size_t ev = 0; ev < event.num_events; ev++) {
           children_queue_.push(single_event);
+          // std::cout << "Enqueued child of priority " << last_priority_ << std::endl;
         }
       }
     }
@@ -91,6 +93,7 @@ public:
     // They are guaranteed to be children of the last dequeued element on uniprocessor
     while(children_queue_.size() > children_priority_queue_.size()) {
       children_priority_queue_.push(last_priority_);
+      // std::cout << "Pushing child of priority " << last_priority_ << std::endl;
     }
 
     // Compare the priority of the top element in the timers queue with the top element in the
@@ -101,12 +104,14 @@ public:
       last_priority_ = timers_queue_.top().priority;
       event = timers_queue_.top().event;
       timers_queue_.pop();
+      // std::cout << "Releasing timer of deadline " << last_priority_ << std::endl;
       return true;
     } else if (!children_queue_.empty()) {
       last_priority_ = children_priority_queue_.top();
       event = children_queue_.top();
       children_queue_.pop();
       children_priority_queue_.pop();
+      // std::cout << "Releasing child of deadline " << last_priority_ << std::endl;
       return true;
     } else {
       return false;
@@ -149,9 +154,9 @@ private:
   // The underlying queue implementation for child subtasks
   std::stack<rclcpp::experimental::executors::ExecutorEvent> children_queue_;
   // Store priority values for elements in children_queue_
-  std::stack<uint32_t> children_priority_queue_;
+  std::stack<int64_t> children_priority_queue_;
   // Priority/period of the last element to be dequeued
-  uint32_t last_priority_ = 0;
+  int64_t last_priority_ = 0;
   // Mutex to protect read/write access to the queue
   mutable std::mutex mutex_;
   // Variable used to notify when an event is added to the queue
